1598162 - Alias buttons under Java 6
In the alias selection panel, when squirrel is just started there is
a selected alias but clicking in the connect button does not connect.
Clicking a second time is needed. Maybe there's a selection lost even
when the alias appears selected.

Caused by a workaround we had in place for a bug in Java 1.2/1.3.


git-svn-id: https://squirrel-sql.svn.sourceforge.net/svnroot/squirrel-sql/trunk@3000 0ceafaf9-a5ae-48a7-aaeb-8ec3e0f2870c