3238037: It's impossible to create a foreign key in oracle 10g, if some tables contains a / in their name eg. recycle bins
For some names of BIN$ objects, Oracle's jdbc driver could not fetch the columns.
Calling DatabaseMetaData#getColumns for a table like BIN$nPl/2NHWRNXgQKjAQgFYEQ==$0 will end in a 
ORA-01424: missing or illegal character following the escape character.

We have to escape the / by //
This is tested with driver version 10.2.0.5.0 an 11.2.0.2.0

git-svn-id: https://squirrel-sql.svn.sourceforge.net/svnroot/squirrel-sql/trunk@6207 0ceafaf9-a5ae-48a7-aaeb-8ec3e0f2870c